updating readmes

+++ App/backend/controllers/uploadController.ts
@@ -7,25 +7,29 @@ const fs = require('fs');
 import * as error from '../errors'
 
 interface ImageRequest extends Request {
-    files?: any | undefined // Define the type of the files property here
+    files?: any | undefined // any lol
 }
 
 export const uploadImage = async (req: ImageRequest, res:Response) => {
-  
     if (!req.files || !req.files.image) {
         throw new error.BadRequestError('Please provide an image file');
     }
 
-    console.log(req.files.image);
-  
+    const allowedMimeTypes = ['image/jpeg', 'image/png'];
+    if (!allowedMimeTypes.includes(req.files.image.mimetype)) {
+        throw new error.BadRequestError('Invalid file type. Only JPEG and PNG files are allowed');
+    }
+
     const result = await cloudinary.uploader.upload(
         req.files.image.tempFilePath, {
             use_filename: true,
             folder: 'csc4330',
+            // transformation: [{ width: 1280, height: 720, crop: 'fill' }], // we can crop to our liking
         }
     );
     fs.unlinkSync(req.files.image.tempFilePath);
 
+    console.log(req.files.image);
     return res.status(StatusCodes.CREATED).json({ image: { src: result.secure_url } });
 };
